backend/requirements.txt
fastapi==0.110.1
flask==3.0.3
httptools==0.6.1
ipykernel==6.29.0
openai==1.17.1
pandas==2.2.2
pip-chill==1.0.3
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
